+++ libshviotqt/src/node/shvnodetree.cpp
@@ -118,7 +118,7 @@ static std::string dump_node(ShvNode *parent, int indent)
 	std::string ret;
 	for(const std::string &pn : parent->childNames()) {
 		ShvNode *chnd = parent->childNode(pn);
-		ret += '\n' + std::string(indent, '\t') + pn;// + ": " + parent->propertyValue(pn).toPrettyString();
+		ret += '\n' + std::string(static_cast<size_t>(indent), '\t') + pn;// + ": " + parent->propertyValue(pn).toPrettyString();
 		if(chnd) {
 			ret += dump_node(chnd, ++indent);
 		}
